Move Selector to own file
